requirements.txt
# Quality Agent — minimal requirements
# Generated by import analysis across all .py files.
# Before: 167 packages (pip freeze). After: 11 packages.

# ── Core AI / API ─────────────────────────────────────────────────────────────
anthropic==0.97.0

# ── Web / API layer ───────────────────────────────────────────────────────────
fastapi==0.136.1
uvicorn==0.43.0
pydantic==2.12.5

# ── Data / ML ─────────────────────────────────────────────────────────────────
pandas==3.0.2
openpyxl==3.1.5          # pandas Excel engine (pd.ExcelFile / pd.read_excel)
SQLAlchemy==2.0.49

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit==1.56.0
plotly==6.7.0
streamlit-google-auth==1.1.8   # Google OAuth login gate

# ── Config / env ──────────────────────────────────────────────────────────────
python-dotenv==1.2.2      # configs/settings.py: from dotenv import load_dotenv

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==9.0.3
